feat: complete app rename to Drift and restore missing inno_setup.iss
