+++ flutter/android/app/src/main/kotlin/com/example/drift/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.app
+package com.example.drift
 
 import io.flutter.embedding.android.FlutterActivity
